Clean up schema command and add JsonSchema derives
