+++ fpx/src/api/handlers/inspect.rs
@@ -1,6 +1,6 @@
 use super::ApiState;
-use crate::api::types;
 use crate::data::libsql::LibSqlStore;
+use crate::schemas;
 use axum::extract::Request;
 use axum::extract::State;
 use axum::response::IntoResponse;
@@ -58,7 +58,7 @@ pub async fn inspect_request_handler(
 
     tx.commit().await.unwrap(); // TODO
 
-    events.broadcast(types::RequestAdded::new(request_id, None).into());
+    events.broadcast(schemas::RequestAdded::new(request_id, None).into());
 
     // TODO: This should return the same payload as the GET /requests/{id} endpoint
     base_url
