+++ fpx/src/api/ws.rs
@@ -1,7 +1,6 @@
-use super::types::{ClientMessage, ServerMessage, FPX_WEBSOCKET_ID_HEADER};
 use super::ApiState;
-use crate::api::types::ServerError;
 use crate::events::ServerEvents;
+use crate::schemas::{ClientMessage, ServerError, ServerMessage, FPX_WEBSOCKET_ID_HEADER};
 use axum::extract::ws::{Message, WebSocket};
 use axum::extract::{State, WebSocketUpgrade};
 use axum::response::Response;
